app/schemas.py
from pydantic import BaseModel, Field


class IngestResponse(BaseModel):
    filename: str
    chunks_indexed: int
    message: str


class ChatRequest(BaseModel):
    question: str = Field(min_length=3, max_length=3000)


class SourceChunk(BaseModel):
    text: str
    score: float


class ChatResponse(BaseModel):
    answer: str
    sources: list[SourceChunk]
