requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
pypdf==5.3.0
scikit-learn==1.6.1
numpy==2.2.3
openai==1.65.2
httpx==0.28.1
python-multipart==0.0.20
pytest==8.3.4
